Use correct field width to accomodate printed headers
We don't actually print the headers we get from ps, we print
uppercase versions of what the user passes in. These can be different
because some can be aliased to others. We were tallying the lengths of
what we got from ps, and the alias the user passed was longer.